feat: Remove 4 byte metadata from bb-produced proof (#13231)
Closes https://github.com/AztecProtocol/barretenberg/issues/1312.

Removes the 4 byte metadata from proof and public inputs produced from
the `to_buffer</*include_size=*/true>()` function. We are able to
serialize and deserialize without this metadata by use many_from_buffer
for deserialization.

The result is that we get to remove a lot of ugly if statements and
weird proof parsing based on the metadata.