fix: deprecate witness circuit size from ECCVM and Translator (#13133)
The circuit sizes in Translator and ECCVM are now given by constexpr
integers. Therefore, the provers do not need to send the circuit sizes
to the verifiers. In its turn, the respective recursive verifiers do not
need to constrain witness `log_circuit_size`s in-circuit.

To support this change, I modified Shplemini interface to accept log
circuit size as an integer as opposed to (unconstrained) witness
`circuit_size`. The derivation of log circuit sizes in other Flavors'
recursive verifiers is still insecure.

Introduced `USE_PADDING` flag in Flavors to allow compile time exclusion
of padding logic in Shplemini and Sumcheck.

Closes https://github.com/AztecProtocol/barretenberg/issues/1040