chore: remove msm sorter (#13668)
The `MsmSorter` class contained logic to sort the inputs to an MSM
{scalars, points} by scalar, sum the points sharing a scalar, then
perform the MSM on the reduced result. This was motivated by the need to
commit to z_perm in the structured setting where the coefficients are
constant over large ranges. Turns out though that since our polys have a
well defined structure, its much better to simply provide the constant
ranges explicitly rather than sorting then adding. Much of the logic of
this class was repurposed into `BatchedAffineAddition` which is what
supports our `commit_structured_with_nonzero_complement` method (used
for `z_perm`).