chore: codeowner and changetest (#13962)
Adds a new test that do some minimal checks to catch if the contract
code have been changed, and add the turtles as codeowners for the test.

That way, we should be able to somewhat avoid not noticing if there are
changes made to the contracts. It only check a few of the contracts;
`rollup`, `governance` and `registry` as those are the big ones and
should deal with most changes.

I only do codeowners on this one file instead of the dir as we don't to
be notified around extra tests or things that would not really cause a
problem if used for nodes.