fix!: fix hiding circuit VK consistency in overflow case (#14011)
Prior to this update, the PG recursive verifier was not treating the
circuit size and log circuit size as witnesses. This led to the
`padding_indicator_array` used in the decider recursive verifier in the
hiding circuit being populated with constant values. This is turn led to
different constraints in the case where the accumulator had a circuit
size larger than the nominal one determined by the structured trace
(which occurs when the overflow mechanism is active for example).

This PR makes those values witnesses and adds an additional CIVC VK
consistency test for the case where one or more of the circuits
overflows and results in an accumulator with larger dyadic size than the
nominal structured trace dyadic size.

---------

Co-authored-by: sergei iakovenko <105737703+iakovenkos@users.noreply.github.com>