feat!: staking escrow (#14339)
Fixes #13870 and the staking portion of #13871.

Starts the work for the staking escrow (not fully tested but more tests
also needed along with the governance part of it, so in a decent spot).

Notable changes:

- as a shared contract there is now the GSE that is escrowing funds
staked, and support moving it along.
- it is another contract that must be called when adding a new rollup if
it is to move along, since it is too dangerous to just move it along in
the hopes that the rollup will work similar to the old one.
- starts the work to have different values for deposit and staying
staked, but atm they are set to the same value to not make this pr even
bigger than it already is
  - see #14304
- deployments in ts are passing along a full "operator" or validator,
e.g., the set of addresses instead of just a single address and using
that for all.
- updates the staking handler to by default follow the canonical.