fix: Attempt tree re-sync after application stop (#14437)
Currently, if the node software is hard stopped, it is possible that the
world state will have inconsistent tree states. Usually this will be the
result of being in the process of committing a block when the
application stopped.

This PR will attempt to bring the trees back in line with each other at
startup.