chore: Try deflake l1-reorg test (#14705)
Try deflaking l1-reorg test `sees new blocks added in an L1 reorg` (see
failing run [here](http://ci.aztec-labs.com/41f6438e922af7f5)), where
the tx that inserted the block after the reorg reverted.

Adding a trace call revealed that it was reverting with an invalid slot
error, probably because we were replaying it in the wrong L1 block. This
PR changes when we pin the L1 block number, which should ensure the tx
lands in the correct L2 slot. It also includes a bunch of unrelated
logging changes to help in other flakes.