chore!: remove num_txs from content commitment (#14947)
The `num_txs` in content commitment is only used to compute the
`min_height` of the wonky l2 to l1 message tree on L1. But it's really
not necessary to have this `min_height`. The leaf of the l2 to l1
message is computed in a fixed way with specific values. It's not
possible to provide a shorter sibling path and a fake leaf that hash to
the correct root.

Removing it allows us to save on calldata and storage cost. Besides, the
number of txs are implied by the blob data, which is committed to by the
`blobs_hash` in content commitment.