fix: Do not use stream ids as unique identifiers (#15054)
Libp2p stream ids are unique _within a muxer_, meaning that for a given
node, if it has connections open to two different peers, each peer may
have a stream with the same id. And since ids are increasing numbers
(1,2,3...), the chances for clashing are high.

This was causing issues in the connection sampler, which stored streams
by id, and closed them by id. So closing a stream after an interaction
with a given peer would accidentally close a different one, leading to
errors.

This PR removes the use of stream ids as identifiers, and just keeps a
set of stream objects in the connection sampler. It also adds a bunch of
logging, and introduces an `P2P_REQRESP_OPTIMISTIC_NEGOTIATION` env var
for toggling libp2p optimistic protocol negotiation on or off (defaults
to disabled, as we use it today).