fix: return tuples from public functions (#15283)
Alternative to
https://github.com/AztecProtocol/aztec-packages/pull/15263 @benesjan

We need to implement Serialize and Deserialize for tuples to return them
from public functions.
This PR generates Serialize and Deserialize impls for an arbitrary tuple
of the given size - currently up to 6-element tuples. `nargo expand` can
be used to look at the generated output although the macro itself is
also commented.

---------

Co-authored-by: benesjan <janbenes1234@gmail.com>
Co-authored-by: Jake Fecher <jake@aztecprotocol.com>