refactor: un-exposing incorrectly exposed functions (#16097)
Fixes #15980

In [this PR](https://github.com/AztecProtocol/aztec-packages/pull/15856)
I incorrectly exposed a couple of functions. In this PR I un-expose
both. The `next_counter` got tackled by me moving the
`emit_contract_class_log` to private context which is where it actually
should be.

For the `call` function I decided to just disable the test funcitons in
which it was used. It's up to the AVM team to re-enable those tests.
Created [an
issue](https://github.com/AztecProtocol/aztec-packages/issues/16099) for
that.