feat: Slash proposers of blocks with invalid attestations (#16147)
Adds a new slasher watcher that listens to invalid-blocks as emitted by
the archiver, and emits wants-to-slash events. Proposers of blocks with
invalid attestations are always slashed, and attestors of blocks _built
on top of blocks with invalid attestations_ are slashed as well. Since
there's no margin for equivocation, these validators are slashed for the
full deposit amount (happy to revisit this, since a bug in the sequencer
could be disastrous).

Also adds a method to the node admin api to inspect the slash payloads
being monitored, used in e2e tests.

Note that, since the slasher client emits the slashing payload
immediately as soon as the watcher emits a want-to-slash, it's possible
that the slash payloads created by this watcher will compete among each
other.