refactor: `RetrievedNote` packing cleanup (#16207)
When trying to use `derive_serialize` in more places I stumbled upon a
`RetrievedNote` where we had a hardcoded `RETRIEVED_NOTE_OVERHEAD`
constant. After deriving the trait impl on the struct I needed to ensure
that it actually matches the real serialized retrieved note. This made
me consider writing a test that would check it matches but when looking
at the use of the constant I figured we could just drop it altogether
when using the newly unblocked Noir syntax of `<RetrievedNote<Note> as
Packable>::N`.

This led to quite a lot of changes but I think the result is worth it as
now the generics on the note oracles are not weird.

Stumbled upon [this](https://github.com/noir-lang/noir/issues/9399)
non-blocking Noir issue when implementing this.