chore: Ensure proving job txs are sorted (#16270)
The txs loaded in the epoch proving job implicitly relied on the txs
being provided in the same order as they are in the block. However, if
tx-collection returned them out of order, then the proving job would try
processing them out of order.

This change ensures the txs are always processed in order.