fix(avm)!: hand-write poseidon2 perm relations (#16452)
Adds a more optimal implementation of the poseidon2 perm relations. The
main goal is to reduce the compilation times on ARM but has also speeds
up `Accumulate` on x86. The next pr will enable it in the AVM

On ARM: saves about 33% of the overall compilation time (~21mins ->
~13mins)
On x86: `accumulate` is ~20% faster ( 43us -> 36us)