feat: Consensus-based slashing proposer contract (#16357)
Implementation for the consensus-based slashing on L1 as discussed
[here](https://linear.app/aztec-labs/document/slashing-discussion-w33-8810b51b11b0#l1-consensus-db4c0919):

> Build a new SlashingProposer version that does not extend from
EmpireBase, but follows same logic on slashing rounds, execution delays,
and allowing votes from the current L2 proposer only.

> Each vote is now an array of proposed slashes. We define a
slashing-unit value, and every slash is a multiple of this unit. Slashes
are represented as a bit vector, where the i-th 4-bits (or 2, or 8,
depending on how much granularity we want) indicate how many
slashing-units proposer i is slashed. An expensive tallying function
computes the effective payload by iterating over all votes.

This still keeps the EmpireSlashProposer around. Next step will be to
let the `Slasher` choose which implementation to use, so we can keep
tests for both, and then rebase this on top of #16284 and tweak the
updated slasher client to use this new SlashProposer.

Other changes:
- Updates the `SlashProposer` to be `Ownable` by governance, who can
execute slash payloads.
- Modifies the `Slasher` deployment due to contract size limits (the
slasher no longer deploys the proposer, instead it needs to be
initialized after deployment with the proposer address).
- Renames `ExtRollupLibN` to have more descriptive names.
- Adds env var config for the new proposer flavor.
- Adds type-checking for L1 contract deployer on ts.