feat!: share transcript between pg and decider (#16642)
Closes https://github.com/AztecProtocol/barretenberg/issues/1453.

Shares the transcript between PG/Merge and the Decider protocols for
hiding kernel soundness. We want to make sure that we can't pick the
accumulator that the Decider runs on without affecting the challenges in
the Decider. This mitigates an origin tag check failure as we had values
from two different transcripts interacting in Decider.