fix: Do not aggregate validator offenses across epochs in a vote (#16695)
When a validator shows up in multiple committees within a round, we used
to aggregate all offenses and vote for them only once. This had side
effects such as that, if the slash amount for each epoch was the max
slashable amount, they would get slashed only once.

This commit changes it so each appearance of a validator in a slashing
round is voted for independently, based on the offenses of that
validator for that specific round.

Fixes
https://linear.app/aztec-labs/issue/TMNT-233/more-clearly-define-expected-behaviour