fix(docs): updating docs structure (#16513)
:warning: this PR builds off of the changes in
https://github.com/AztecProtocol/aztec-packages/pull/16442, so merge
that one first

This PR makes several edits to the docs. The main impetus behind the
changes being made is to make a more linear learning journey for
developers as they move down the sidebar.

- moves the guides section above the tutorials
- flattens the pages inside the "Developer smart contracts" section
- makes the "developing smart contracts" section a more linear flow for
how a developer should be moving through this content
- moves some of the technical explanations to the concepts section
- adds a several pages to the Guides section including:
  - project sturcturing
  - data types
  - custom types
  - defining functions (initializer page moved into here)
- removes some explainers and reference about storage from the
references section into the guides section

closes: https://github.com/AztecProtocol/dev-rel/issues/598