feat!: The Great Wallet Refactor (#16865)
Second part of the `Wallet` refactor that moves us closer to a fully
fledged `Wallet SDK` and PXEless apps. Focuses on completely unbinding
the `Wallet` interface from a particular `AztecAddress`. The former now
refers to the wallet *software*, which exposes a clear API to apps. This
API is not yet finalized!

`ContractFunctionInteraction` and friends have been put on a diet, and
most of their responsibilities are now the Wallet's

As a consequence of this PR, every single e2e test, the playground, the
cli-wallet and boxes have been altered. It's not pretty (yet), but it
will be, I promise.