refactor!: dropping PXE JSON RPC (#17228)
PXE is becoming just a lib used by the wallet and for this reason I am
dropping the corresponding JSON RPC Server.

This leads to a large refactor of basically all the setups that dealt
with PXE and wallets.

This is a followup work of [this
PR](https://github.com/AztecProtocol/aztec-packages/pull/17163) in which
I had this tasks planned:
1. Drop PXE JSON RPC server,
2. have Wallet instantiate PXE instead of having it passed in as an arg
to a constructor,
3. revisit the PXE methods exposed on the TestWallet.

In this PR I tackled points 1 and 2.