chore: refactor and add tests for tx base rollups (#17259)
Restructured code, renamed some variables, and added more tests.

Fixed 2 bugs:
- The constant for the max blob fields of a tx was wrong. It only
considered 1 extra field per contract class log, but we added 2 fields
(address and log length) to the blobs per log. The circuit would fail if
a tx had a full tx effect.
- The check for the lengths of private logs and contract class logs was
in the unconstrained function that serialized tx effect for adding to
blob. I've moved it to private tail since it's not too expensive and
could avoid having to check in both the tx base and the sequencer.

Other notable changes:
- Validate the vk hash is in vk tree in simulated circuits. Previously
this and `verify_proof` were disabled in unconstrained mode.
- Remove redundant code path for indexed tree insertion. We used general
utility functions to insert nullifiers and public data write. In the
utils they checked that the low leaf could be equal to or less than the
new leaf, and updated the low leaf or inserted the new leaf accordingly.
- For nullifiers, the new leaf can't be the same as the low leaf. So now
we use a specific function `batch_insert_no_update` that only allows
inserting new leaves.
- For public data write in private tx base, it can only be updating the
fee payer's balance. Now we have a dedicated function to simply update
an existing leaf in the tree.