chore: change leaf index to Field (#17352)
The `next_available_leaf_index` in `AppendOnlyTreeSnapshot` was `u32`.
The `leaf_index` for membership check is `Field`. And the trees can be
up to 40 deep, so the leaf indices need to be at least `u64`.
Changing the `next_available_leaf_index` in `AppendOnlyTreeSnapshot` to
`Field` to allow trees to grow fully.
Choosing Field instead of `u64` because of its cheaper `to_le_bits`
operation.