feat(avm): flatten flavor entities (#17363)
Makes `AllEntities` of the AVM flavor be a flat array instead of a
struct with 3.5k+ named members.

In the AVM we access members using `.get(COLUMN_ID)` and not `.column`.
This was done some time ago to enable a speedup for `get_row` and also
re-used to enable a speedup on edge extension for sumcheck.

Why this change?
* Cleaner code
* Column access is simpler (before I needed a statically cached table of
pointers).
* `get_all()` and friends can now return spans. This is not major, since
creating a span is very cheap and we were apparently spending a lot of
time creating the RefArrays and RefVectors. I don't expect a big
performance impact though, since we should've already eliminated any use
of `get_*()` in hot loops. Note that whenever BB requires a `RefVector`,
I'm converting with `RefVector<T>::from_span`.
* Should make compilation time a bit faster

TESTS: I had to migrate tests to not use AvmFullRow, because I changed
the ordering and it's a pain to migrate.

Resolves #17288.