feat: TXE dispatch function name zod validation check (#17426)
We didn't perform a runtime check of the function name in the process
function of dispatcher. In this PR I fix that.

Without this PR the handler would most likely fail with some random
error. With this PR they get a nice error message that there is no
oracle handler corresponding to the `functionName`.

Also sneaked in a fix of the `TXEOracleFunctionName` type.