refactor: tagging related cleanup (#17451)
I started working on improving unconstrained tagging as proposed in
[this forum
post](https://forum.aztec.network/t/on-note-discovery-and-index-coordination/7165)
and the initial step in implementing that is the following:

1. store the tagging indexes requested during execution in a cache (just
like we do with notes),
2. when the execution is finished take the cache and along with a tx
hash dump it into db.

Given that the cache will reside in `PrivateExecutionOracle` I now need
to orchestrate the sync there.

When trying to do that (to be done in a [PR up the
stack](https://github.com/AztecProtocol/aztec-packages/pull/17445)) I
realized that the current tagging-related API was cumbersome and this
made me do the following in this PR:
1. Have the API accept directionalAppTaggingSecret on the input instead
of appTaggingSecret and recipient as now I can just compute it in the
`PrivateExecutionOracle` and request info based on that,
2. I moved the tagging functionality from stdlib to PXE as there was no
reason to have it in stdlib,
3. I simplified a bunch of random things.

This PR is a bit random so if something is not clear LMK.