refactor: returning last used tagging index (#17534)
As
[discussed](https://github.com/AztecProtocol/aztec-packages/pull/17445#discussion_r2405913527)
in a PR down the stack returning the next tagging index in various
places of tagging was confusing and it's better to return the last used
one. In this PR I do this change.

Some of the code is quite weird but I think it doesn't make sense to
invest time in that as I will replace all the tagging sync it all in
followup PRs.