refactor!: introducing `#[external(...)]` macro (#17827)
Closes
https://linear.app/aztec-labs/issue/F-19/rename-privateorpublicorutility-to-externalprivateorpublicorutility

Having `#[private]` and `#[public]` macros was misleading because in
Solidity these define function visibility while in Aztec's case they
define the environment in which the functions are executed. For this
reason I introduce `#[external(...)]` macro while the argument of the
attribute defines the execution environment: `private`, `public` or
`utility`.

## Notes for reviewer
- I did a refactor in a [PR up the
stack](https://github.com/AztecProtocol/aztec-packages/pull/17840) which
allowed me to drop a lot of the complexity introduced in this PR. For
this reason I encourage you to check that one first to have context on
what in this PR is a problem and what is not. It might make sense to
just merge these 2 PRs into 1 but I think that would make it all harder
to review.
- This PR touches a lot of files but all the juicy bits are in the
`noir-projects/aztec-nr/aztec/src/macros/` directory so I encourage you
to just check that first. The `noir-contracts` dir can be completely
ignored as there I just update the imports and the macros.
- The new naming is now in conflict with the `#[internal]` macro as
having both external and internal applied to a function is valid:

<img width="874" height="182" alt="image"
src="https://github.com/user-attachments/assets/4582a51a-225c-4a4c-9b9a-d2c2d413d927"
/>

This issue will be tackled in the near future.