feat: constructor arg to delay reward claiming (#17890)
Fix TMNT-438

Edit: (by @LHerskind)

The code had issues with naming convention and code size. To address the
code size issue. I had to move more stuff into the `RewardLib` and then
use it as an external library to save space.
The diff is a bit larger than i would like because of that. But it meant
that it will at least fit.

I moved the deployment of the booster into this external reward library.

I did also update the evm version and fixed a `transfer` to
`safeTransfer` in the token portal example.

Note that the base branch should be changed to next as well, since I
think we should skip the governance to slash pr as it already exists.