feat(ci): flake_error_threshold kludge (#17962)
Due to special circumstances with regard to known instability, we will
be setting a flake threshold on this group of tests due to high
correlation in their failures. We will revisit this as priorities allow.

This forms only weak precedent for doing this in the future. Make sure
to discuss any reuse of this feature - it is preferable to not have rest
of team merge for 2 days rather than destabilize. In this circumstance,
we felt factors weighed against trying to stabilize at all costs.