feat!: remove context from NoteEmission methods (#17024)
In discussing how to improve developer experience by getting rid of the
`context` object as much as possible, I realized some of it is quite low
hanging fruit. We can take a similar approach to the `storage` object
and have the aztec-nr intermediate structs keep references to the
context, resulting in the end-user invoking methods that already
reference the context instead of having to manually provide it. The diff
speaks for itself.

Closes
https://linear.app/aztec-labs/issue/F-109/remove-context-when-delivering-messages.