fix(aztec-nr): Avoid relying on unquoted integers changing types (#18299)
After https://github.com/noir-lang/noir/pull/10330, noir will start
enforcing the types of unquoted integers remain the same type. Before,
when you unquoted an integer such as `let n: u32 = 4;` e.g. in `quote {
let x = $n; }` the unquoted `n` would be a polymorphic integer literal
of any type. With the above PR, it unquotes instead to `let x = 4u32;` -
the type is preserved. We consider it a bug to rely on the previous
behavior, and there were 3 instances of it in aztec-nr so this is a
patch for those cases.