chore: encode block header data in blob (#18312)
Currently the block header can still be reconstructed from the calldata
submitted to L1 when proposing a block/checkpoint. But once there's more
than 1 block in a checkpoint, the values for individual blocks would no
longer be available from the calldata.

This pr encodes the following fields from each block header in blobs, so
that we can rebuild the block headers without having to build the trees
or look up from other sources:
- blockNumber
- timestamp
- totalManaUsed
- lastArchiveRoot
- stateReference:
  - noteHash root and nextAvailableLeafIndex
  - nullifier root and nextAvailableLeafIndex
  - publicData root and nextAvailableLeafIndex
  - l1toL2Msg root and nextAvailableLeafIndex

The `stateReference` is still part of the calldata that is signed over
by the committee. Note that this is not checked against the proof. If
half the committee signs over a wrong `stateReference`, the epoch can
still be proven. And the block building can still proceed as the real
`stateReference` and the tx effects are encoded in the blob data, and
that's where the archiver gets the data from.
Without `stateReference` in the signature digest, the committee still
agrees to the new state by signing over the last archive and the blobs
hash. But keeping it as the calldata for now as it might serve as a good
training wheel.