fix: merkle_check_trace test passes assertion that final index is 0 (#18487)
tracegen asserts that the final index (after halving for each layer in
the sibling path) is 0, but the test was starting with (for example)
index 6 with a sibling path of length 1, so tracegen was only halving
once and ending with index of 3. This PR just changes the test so that
sibling paths align with indices such that tracegen halves them enough
times to reach 0.