chore: granular ethereum imports (#18860)
As title. A push to fix cli startup times. This is one of several import
granularization efforts. At the end of it all we can at least print the
cli help in 0.5s.

I propose the following:
* We **remove barrel imports**. So no more `index.ts` that's just a list
of `export * from 'foo'`.
* In any project, there are files that are deemed part of the public api
(they were exported directly or indirectly via the barrel import), and
there are files that are not deemed part of the public api.
* Any file that is deemed part of the public api is exported directly
through the `package.json` `export` api. If it's good enough to be
exported through a barrel, it's good enough to have it's own export on
the package.
* This will make it *explicit and clear* what we're presenting as the
packages public api.
* It makes it easier to trivially time the importation of each
entrypoint for debugging slow stuff.
* It makes it easier to see messy apis and take action to clean them up.