fix: make pxe check tx hash on note storage (#18900)
Part of a broader set of changes re. to note and event storage. I
changed note storage so that it checks that the note hash exists in the
effects of the tx that we queried for, and adjusted tests accordingly
(mirroring the work done in the `deliverEvent` tests by @mverzilli). I
did not remove the rest of the behavior during validation, e.g. checking
tree leaves, as that will be done in follow up PRs as part of e.g.
F-218.