refactor(PXE): clean up block sync (#19026)
Block sync is currently handled through collaboration of `Synchronizer`,
`SyncDataProvider`, and even `PXEOracleInterface#getAnchorBlock`.

## Current problems
- The names aren't very intuitive (`Synchronizer` and `SyncDataProvider`
only handle chain state data at the end of the day).
- Any code running inside `ContractFunctionSimulator` can call
`getAnchorBlock` whenever it pleases. This works currently because calls
to `Synchronizer` are carefully orchestrated from PXE. But it seems
error prone, obscure, and it gives the false impression that function
simulation happens on live block data when it actually depends on a
static anchor.
- There are multiple ways throughout the app to get the anchor block:
`Synchronizer#getSynchedBlock` (unused),
`PXEOracleInterface#getAnchorBlock` (unnecessary, inappropriate),
`SyncDataProvider#getBlockNumber` (redundant), and
`SyncDataProvider#getBlockHeader` (the one that should be the source of
truth)
- Many checks are relying on block number when they should rely on block
hash.

## Refactor

(note: adjusted naming after discussing with @benesjan)

- Rename `Synchronizer` => `BlockSynchronizer`
- Rename `SyncDataProvider` => `AnchorBlockDataProvider`
- Remove `AnchorBlockDataProvider#getBlockNumber`.
`AnchorBlockDataProvider#getBlockHeader` remains the only way to consume
current block information for the rest of the app. To get the block
number, there's a simple `getBlockNumber` accessor in `BlockHeader`.
- Remove `ExecutionDataProvider#getAnchorBlock`. Instead, explicitly
pass anchor block headers as params to functions in "the realm" of
`ContractFunctionSimulator` (`run`, `runUtility`,
`varifyCurrentClassId`, etc).

## Out of scope of PR

> - Many checks are relying on block number when they should rely on
block hash.

This should be tackled on a per use case/feature basis in future PRs,
though relying on `getBlockHeader` solely seems aligned in the right
direction.

Closes F-225